+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
-## [0.0.1] - 2019-10-09
+
+## [1.0.2] - 2019-10-10
+### Added
+- Disable creation of settings file when there are no excluded files
+- Only show the statusbar toggle when there are items hidden
+- Option to switch the statusbar toggle between always active or only when items are hidden
+
+## [1.0.1] - 2019-10-10
 ### Added
 - Support to hide files
 - Support to hide folders